LayoutTests/fast/hidpi/image-srcset-invalid-descriptor-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("foo").clientWidth==100 is true
PASS document.getElementById("foo2").clientWidth==200 is true
This test passes if the image below is a 100px blue square when the deviceScaleFactor is 1.
It tests that even though the 1x resource contains many invalid descriptors, only the invalid descriptors are ignored (according to the spec).
  
